Moved socket_interface.py to utilities
