+++ openwpm/mp_logger.py
@@ -16,8 +16,9 @@ from multiprocess import JoinableQueue
 from sentry_sdk.integrations.logging import BreadcrumbHandler, EventHandler
 from tblib import pickling_support
 
+from openwpm.utilities.socket_interface import ServerSocket
+
 from .commands.utils.webdriver_utils import parse_neterror
-from .socket_interface import ServerSocket
 
 pickling_support.install()
